Drop live schema support.
