+++ modules/client/src/test/java/org/apache/ignite/client/fakes/FakeInternalTable.java
@@ -30,7 +30,6 @@ import org.apache.ignite.internal.storage.engine.TableStorage;
 import org.apache.ignite.internal.table.InternalTable;
 import org.apache.ignite.lang.IgniteInternalException;
 import org.apache.ignite.lang.IgniteUuid;
-import org.apache.ignite.schema.definition.SchemaManagementMode;
 import org.apache.ignite.tx.Transaction;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -81,18 +80,6 @@ public class FakeInternalTable implements InternalTable {
     @Override
     public @NotNull String tableName() {
         return tableName;
-    }
-    
-    /** {@inheritDoc} */
-    @Override
-    public @NotNull SchemaManagementMode schemaMode() {
-        return SchemaManagementMode.STRICT;
-    }
-    
-    /** {@inheritDoc} */
-    @Override
-    public void schema(SchemaManagementMode schemaMode) {
-        // No-op.
     }
     
     /** {@inheritDoc} */
