+++ modules/runner/src/integrationTest/java/org/apache/ignite/internal/runner/app/ItSchemaChangeTableViewTest.java
@@ -30,6 +30,7 @@ import org.apache.ignite.schema.SchemaBuilders;
 import org.apache.ignite.schema.definition.ColumnDefinition;
 import org.apache.ignite.schema.definition.ColumnType;
 import org.apache.ignite.table.RecordView;
+import org.apache.ignite.table.Table;
 import org.apache.ignite.table.Tuple;
 import org.junit.jupiter.api.Test;
 
@@ -43,32 +44,32 @@ class ItSchemaChangeTableViewTest extends AbstractSchemaChangeTest {
     @Test
     public void testDropColumn() {
         List<Ignite> grid = startGrid();
-    
+        
         createTable(grid);
-    
+        
         RecordView<Tuple> tbl = grid.get(0).tables().table(TABLE).recordView();
-    
+        
         tbl.insert(Tuple.create().set("key", 1L).set("valInt", 111).set("valStr", "str"));
-    
+        
         dropColumn(grid, "valStr");
-    
+        
         // Check old row conversion.
         final Tuple keyTuple = Tuple.create().set("key", 1L);
-    
+        
         assertEquals(1, (Long) tbl.get(keyTuple).value("key"));
         assertEquals(111, (Integer) tbl.get(keyTuple).value("valInt"));
         assertThrows(IllegalArgumentException.class, () -> tbl.get(keyTuple).value("valStr"));
-    
+        
         // Check tuple of outdated schema.
         assertThrowsWithCause(SchemaMismatchException.class,
                 () -> tbl.insert(Tuple.create().set("key", 2L).set("valInt", -222).set("valStr", "str"))
         );
-    
+        
         // Check tuple of correct schema.
         tbl.insert(Tuple.create().set("key", 2L).set("valInt", 222));
-    
+        
         final Tuple keyTuple2 = Tuple.create().set("key", 2L);
-    
+        
         assertEquals(2, (Long) tbl.get(keyTuple2).value("key"));
         assertEquals(222, (Integer) tbl.get(keyTuple2).value("valInt"));
         assertThrows(IllegalArgumentException.class, () -> tbl.get(keyTuple2).value("valStr"));
@@ -82,11 +83,11 @@ class ItSchemaChangeTableViewTest extends AbstractSchemaChangeTest {
         List<Ignite> grid = startGrid();
         
         createTable(grid);
-    
+        
         RecordView<Tuple> tbl = grid.get(0).tables().table(TABLE).recordView();
-    
+        
         tbl.insert(Tuple.create().set("key", 1L).set("valInt", 111));
-    
+        
         assertThrowsWithCause(SchemaMismatchException.class,
                 () -> tbl.insert(Tuple.create().set("key", 1L).set("valInt", -111).set("valStrNew", "str"))
         );
@@ -118,34 +119,34 @@ class ItSchemaChangeTableViewTest extends AbstractSchemaChangeTest {
         List<Ignite> grid = startGrid();
         
         createTable(grid);
-    
+        
         RecordView<Tuple> tbl = grid.get(0).tables().table(TABLE).recordView();
-    
+        
         tbl.insert(Tuple.create().set("key", 1L).set("valInt", 111));
-    
+        
         assertThrowsWithCause(SchemaMismatchException.class,
                 () -> tbl.insert(Tuple.create().set("key", 2L).set("valRenamed", -222))
         );
-    
+        
         renameColumn(grid, "valInt", "valRenamed");
-    
+        
         // Check old row conversion.
         Tuple keyTuple1 = Tuple.create().set("key", 1L);
-    
+        
         assertEquals(1, (Long) tbl.get(keyTuple1).value("key"));
         assertEquals(111, (Integer) tbl.get(keyTuple1).value("valRenamed"));
         assertThrows(IllegalArgumentException.class, () -> tbl.get(keyTuple1).value("valInt"));
-    
+        
         // Check tuple of outdated schema.
         assertThrowsWithCause(SchemaMismatchException.class,
                 () -> tbl.insert(Tuple.create().set("key", 2L).set("valInt", -222))
         );
-    
+        
         // Check tuple of correct schema.
         tbl.insert(Tuple.create().set("key", 2L).set("valRenamed", 222));
-    
+        
         Tuple keyTuple2 = Tuple.create().set("key", 2L);
-    
+        
         assertEquals(2, (Long) tbl.get(keyTuple2).value("key"));
         assertEquals(222, (Integer) tbl.get(keyTuple2).value("valRenamed"));
         assertThrows(IllegalArgumentException.class, () -> tbl.get(keyTuple2).value("valInt"));
@@ -159,33 +160,33 @@ class ItSchemaChangeTableViewTest extends AbstractSchemaChangeTest {
         List<Ignite> grid = startGrid();
         
         createTable(grid);
-    
+        
         RecordView<Tuple> tbl = grid.get(0).tables().table(TABLE).recordView();
-    
+        
         tbl.insert(Tuple.create().set("key", 1L).set("valInt", 111));
-    
+        
         assertThrowsWithCause(SchemaMismatchException.class,
                 () -> tbl.insert(Tuple.create().set("key", 2L).set("val2", -222))
         );
-    
+        
         renameColumn(grid, "valInt", "val2");
         addColumn(grid, SchemaBuilders.column("valInt", ColumnType.INT32).asNullable().withDefaultValueExpression(-1).build());
-    
+        
         // Check old row conversion.
         Tuple keyTuple1 = Tuple.create().set("key", 1L);
-    
+        
         assertEquals(1, (Long) tbl.get(keyTuple1).value("key"));
         assertEquals(111, (Integer) tbl.get(keyTuple1).value("val2"));
         assertEquals(-1, (Integer) tbl.get(keyTuple1).value("valInt"));
-    
+        
         // Check tuple of outdated schema.
         assertNull(tbl.get(Tuple.create().set("key", 2L)));
-    
+        
         // Check tuple of correct schema.
         tbl.insert(Tuple.create().set("key", 2L).set("val2", 222));
-    
+        
         Tuple keyTuple2 = Tuple.create().set("key", 2L);
-    
+        
         assertEquals(2, (Long) tbl.get(keyTuple2).value("key"));
         assertEquals(222, (Integer) tbl.get(keyTuple2).value("val2"));
         assertEquals(-1, (Integer) tbl.get(keyTuple2).value("valInt"));
@@ -204,56 +205,56 @@ class ItSchemaChangeTableViewTest extends AbstractSchemaChangeTest {
                 .build();
         
         RecordView<Tuple> tbl = grid.get(0).tables().table(TABLE).recordView();
-    
+        
         tbl.insert(Tuple.create().set("key", 1L).set("valInt", 111));
-    
+        
         assertThrowsWithCause(SchemaMismatchException.class, () -> tbl.insert(
                 Tuple.create().set("key", 2L).set("val", "I'not exists"))
         );
         
         addColumn(grid, column);
-    
+        
         assertNull(tbl.get(Tuple.create().set("key", 2L)));
-    
+        
         tbl.insert(Tuple.create().set("key", 2L).set("valInt", 222).set("val", "string"));
-    
+        
         tbl.insert(Tuple.create().set("key", 3L).set("valInt", 333));
         
         dropColumn(grid, column.name());
-    
+        
         tbl.insert(Tuple.create().set("key", 4L).set("valInt", 444));
-    
+        
         assertThrowsWithCause(SchemaMismatchException.class, () -> tbl.insert(
                 Tuple.create().set("key", 4L).set("val", "I'm not exist"))
         );
-    
+        
         addColumn(grid, SchemaBuilders.column("val", ColumnType.string()).withDefaultValueExpression("default").build());
-    
+        
         tbl.insert(Tuple.create().set("key", 5L).set("valInt", 555));
-    
+        
         // Check old row conversion.
         Tuple keyTuple1 = Tuple.create().set("key", 1L);
-    
+        
         assertEquals(111, (Integer) tbl.get(keyTuple1).value("valInt"));
         assertEquals("default", tbl.get(keyTuple1).value("val"));
-    
+        
         Tuple keyTuple2 = Tuple.create().set("key", 2L);
-    
+        
         assertEquals(222, (Integer) tbl.get(keyTuple2).value("valInt"));
         assertEquals("default", tbl.get(keyTuple2).value("val"));
-    
+        
         Tuple keyTuple3 = Tuple.create().set("key", 3L);
-    
+        
         assertEquals(333, (Integer) tbl.get(keyTuple3).value("valInt"));
         assertEquals("default", tbl.get(keyTuple3).value("val"));
-    
+        
         Tuple keyTuple4 = Tuple.create().set("key", 4L);
-    
+        
         assertEquals(444, (Integer) tbl.get(keyTuple4).value("valInt"));
         assertEquals("default", tbl.get(keyTuple4).value("val"));
-    
+        
         Tuple keyTuple5 = Tuple.create().set("key", 5L);
-    
+        
         assertEquals(555, (Integer) tbl.get(keyTuple5).value("valInt"));
         assertEquals("default", tbl.get(keyTuple5).value("val"));
     }
@@ -270,36 +271,52 @@ class ItSchemaChangeTableViewTest extends AbstractSchemaChangeTest {
         RecordView<Tuple> tbl = grid.get(0).tables().table(TABLE).recordView();
         
         final String colName = "valStr";
-    
+        
         tbl.insert(Tuple.create().set("key", 1L).set("valInt", 111));
         
         changeDefault(grid, colName, (Supplier<Object> & Serializable) () -> "newDefault");
         addColumn(grid, SchemaBuilders.column("val", ColumnType.string()).withDefaultValueExpression("newDefault").build());
-    
+        
         tbl.insert(Tuple.create().set("key", 2L).set("valInt", 222));
-    
+        
         changeDefault(grid, colName, (Supplier<Object> & Serializable) () -> "brandNewDefault");
         changeDefault(grid, "val", (Supplier<Object> & Serializable) () -> "brandNewDefault");
-    
+        
         tbl.insert(Tuple.create().set("key", 3L).set("valInt", 333));
-    
+        
         // Check old row conversion.
         Tuple keyTuple1 = Tuple.create().set("key", 1L);
-    
+        
         assertEquals(111, (Integer) tbl.get(keyTuple1).value("valInt"));
         assertEquals("default", tbl.get(keyTuple1).value("valStr"));
         assertEquals("newDefault", tbl.get(keyTuple1).value("val"));
-    
+        
         Tuple keyTuple2 = Tuple.create().set("key", 2L);
-    
+        
         assertEquals(222, (Integer) tbl.get(keyTuple2).value("valInt"));
         assertEquals("newDefault", tbl.get(keyTuple2).value("valStr"));
         assertEquals("newDefault", tbl.get(keyTuple2).value("val"));
-    
+        
         Tuple keyTuple3 = Tuple.create().set("key", 3L);
-    
+        
         assertEquals(333, (Integer) tbl.get(keyTuple3).value("valInt"));
         assertEquals("brandNewDefault", tbl.get(keyTuple3).value("valStr"));
         assertEquals("brandNewDefault", tbl.get(keyTuple3).value("val"));
+    }
+    
+    /**
+     * Check operation failed if unknown column found.
+     */
+    @Test
+    public void testStrictSchemaInsertRowOfNewSchema() {
+        List<Ignite> grid = startGrid();
+        
+        createTable(grid);
+        
+        Table tbl = grid.get(0).tables().table(TABLE);
+        
+        Tuple tuple = Tuple.create().set("key", 1L).set("unknownColumn", 10);
+        
+        assertThrowsWithCause(SchemaMismatchException.class, () -> tbl.recordView().insert(tuple));
     }
 }
