+++ modules/table/src/main/java/org/apache/ignite/internal/table/InternalTable.java
@@ -24,7 +24,6 @@ import java.util.concurrent.Flow.Publisher;
 import org.apache.ignite.internal.schema.BinaryRow;
 import org.apache.ignite.internal.storage.engine.TableStorage;
 import org.apache.ignite.lang.IgniteUuid;
-import org.apache.ignite.schema.definition.SchemaManagementMode;
 import org.apache.ignite.tx.Transaction;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -54,18 +53,6 @@ public interface InternalTable extends AutoCloseable {
      * @return Table name.
      */
     @NotNull String tableName();
-
-    /**
-     * Gets a schema mode of the table.
-     *
-     * @return Schema mode.
-     */
-    @NotNull SchemaManagementMode schemaMode();
-
-    /**
-     * Sets schema mode for the table.
-     */
-    void schema(SchemaManagementMode schemaMode);
 
     /**
      * Asynchronously gets a row with same key columns values as given one from the table.
