+++ modules/table/src/main/java/org/apache/ignite/internal/table/RecordBinaryViewImpl.java
@@ -59,7 +59,7 @@ public class RecordBinaryViewImpl extends AbstractTableView implements RecordVie
     public RecordBinaryViewImpl(InternalTable tbl, SchemaRegistry schemaReg, TableManager tblMgr, @Nullable Transaction tx) {
         super(tbl, schemaReg, tx);
         
-        marsh = new TupleMarshallerImpl(tblMgr, tbl, schemaReg);
+        marsh = new TupleMarshallerImpl(schemaReg);
         
         this.tblMgr = tblMgr;
     }
