+++ modules/table/src/main/java/org/apache/ignite/internal/table/TableImpl.java
@@ -20,7 +20,6 @@ package org.apache.ignite.internal.table;
 import org.apache.ignite.internal.schema.SchemaRegistry;
 import org.apache.ignite.internal.table.distributed.TableManager;
 import org.apache.ignite.lang.IgniteUuid;
-import org.apache.ignite.schema.definition.SchemaManagementMode;
 import org.apache.ignite.table.KeyValueView;
 import org.apache.ignite.table.RecordView;
 import org.apache.ignite.table.Table;
@@ -100,14 +99,5 @@ public class TableImpl implements Table {
     /** {@inheritDoc} */
     @Override public KeyValueView<Tuple, Tuple> keyValueView() {
         return new KeyValueBinaryViewImpl(tbl, schemaReg, tblMgr, null);
-    }
-
-    /**
-     * Sets new schema management mode.
-     *
-     * @param schemaMode New schema management mode.
-     */
-    public void schemaMode(SchemaManagementMode schemaMode) {
-        this.tbl.schema(schemaMode);
     }
 }
