+++ modules/table/src/main/java/org/apache/ignite/internal/table/distributed/storage/InternalTableImpl.java
@@ -63,7 +63,6 @@ import org.apache.ignite.lang.LoggerMessageHelper;
 import org.apache.ignite.network.NetworkAddress;
 import org.apache.ignite.raft.client.Peer;
 import org.apache.ignite.raft.client.service.RaftGroupService;
-import org.apache.ignite.schema.definition.SchemaManagementMode;
 import org.apache.ignite.tx.Transaction;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -94,9 +93,6 @@ public class InternalTableImpl implements InternalTable {
     /** Resolver that resolves a network address to node id. */
     private final Function<NetworkAddress, String> netAddrResolver;
 
-    /** Table schema mode. */
-    private volatile SchemaManagementMode schemaMode;
-
     /** Storage for table data. */
     private final TableStorage tableStorage;
 
@@ -122,8 +118,6 @@ public class InternalTableImpl implements InternalTable {
         this.partitions = partitions;
         this.netAddrResolver = netAddrResolver;
         this.tableStorage = tableStorage;
-
-        this.schemaMode = SchemaManagementMode.STRICT;
     }
 
     /** {@inheritDoc} */
@@ -148,18 +142,6 @@ public class InternalTableImpl implements InternalTable {
     @Override
     public String tableName() {
         return tableName;
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    public SchemaManagementMode schemaMode() {
-        return schemaMode;
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    public void schema(SchemaManagementMode schemaMode) {
-        this.schemaMode = schemaMode;
     }
 
     /** {@inheritDoc} */
