+++ modules/table/src/test/java/org/apache/ignite/internal/benchmarks/TupleMarshallerFixlenOnlyBenchmark.java
@@ -105,7 +105,7 @@ public class TupleMarshallerFixlenOnlyBenchmark {
                         .toArray(Column[]::new)
         );
 
-        marshaller = new TupleMarshallerImpl(null, null, new SchemaRegistryImpl(v -> null) {
+        marshaller = new TupleMarshallerImpl(new SchemaRegistryImpl(v -> null) {
             @Override
             public SchemaDescriptor schema() {
                 return schema;
