+++ modules/table/src/test/java/org/apache/ignite/internal/benchmarks/TupleMarshallerVarlenOnlyBenchmark.java
@@ -119,7 +119,7 @@ public class TupleMarshallerVarlenOnlyBenchmark {
                         .toArray(Column[]::new)
         );
 
-        marshaller = new TupleMarshallerImpl(null, null, new SchemaRegistryImpl(v -> null) {
+        marshaller = new TupleMarshallerImpl(new SchemaRegistryImpl(v -> null) {
             @Override
             public SchemaDescriptor schema() {
                 return schema;
