+++ modules/table/src/test/java/org/apache/ignite/internal/table/SchemaValidationTest.java
@@ -35,9 +35,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.function.Executable;
 
 /**
- * Check data by strict schema.
+ * Checks if data compliant with the schema, otherwise the correct exception is thrown.
  */
-public class StrictSchemaOperationsTest {
+public class SchemaValidationTest {
     @Test
     public void columnNotExist() {
         SchemaDescriptor schema = new SchemaDescriptor(
@@ -50,7 +50,7 @@ public class StrictSchemaOperationsTest {
         
         assertThrowsWithCause(SchemaMismatchException.class, () -> recView.insert(Tuple.create().set("id", 0L).set("invalidCol", 0)));
     }
-
+    
     @Test
     public void schemaMismatch() {
         SchemaDescriptor schema = new SchemaDescriptor(
@@ -78,7 +78,7 @@ public class StrictSchemaOperationsTest {
         assertThrowsWithCause(SchemaMismatchException.class, () -> tbl.keyValueView().put(Tuple.create().set("id", 0L).set("affId", 1L),
                 Tuple.create().set("id", 0L).set("val", 0L)));
     }
-
+    
     @Test
     public void typeMismatch() {
         SchemaDescriptor schema = new SchemaDescriptor(
@@ -102,7 +102,7 @@ public class StrictSchemaOperationsTest {
         assertThrowsWithCause(InvalidTypeException.class,
                 () -> tbl.insert(Tuple.create().set("id", 0L).set("valBytes", new byte[]{0, 1, 2, 3})));
     }
-
+    
     @Test
     public void stringTypeMatch() {
         SchemaDescriptor schema = new SchemaDescriptor(
@@ -126,7 +126,7 @@ public class StrictSchemaOperationsTest {
         // Check string 3 char length and 9 bytes.
         tbl.insert(tuple.set("valString", "我是谁"));
     }
-
+    
     @Test
     public void bytesTypeMatch() {
         SchemaDescriptor schema = new SchemaDescriptor(
